+++ includes/elementor/elementor-tag-packageratestable.php
@@ -74,6 +74,7 @@ class Elementor_Tag_PackageRatesTable extends \Elementor\Core\DynamicTags\Tag {
 				$accommodation_resaweburl = esc_html(get_field('resaweb_url', $accommodation->ID));
 
 				$defaultnights = $triptype_defaultnights;
+				echo '*def01-'.$defaultnights;
 				$accommodation_type = get_field('accommodation_type', $accommodation->ID); // term object accommodation_type
 
 				if(!empty($accommodation_type)){
@@ -81,17 +82,20 @@ class Elementor_Tag_PackageRatesTable extends \Elementor\Core\DynamicTags\Tag {
 					$accommodation_type_for_trip_type = get_field( 'accommodation_type', 'trip_type_'.$triptype->term_id);
 					// Trip Type is not related to accommodation type, exclude
 					if(!(is_array($accommodation_type_for_trip_type) && in_array($accommodation_type->term_id, $accommodation_type_for_trip_type))){
-						echo 'remove';
 						continue;
 					}
 
 					$accommodation_defaultnights = get_field('defaultnights', $accommodation_type->taxonomy . '_' . $accommodation_type->term_id);
+					echo $accommodation_type->taxonomy . '_' . $accommodation_type->term_id;
+					echo '*$accommodation_defaultnights:'.$accommodation_defaultnights;
 					if(!empty($accommodation_defaultnights)){
 						$defaultnights = $accommodation_defaultnights;
+						echo '*def02-'.$defaultnights;
 					}
 				}
 
 				if(!in_array($defaultnights, $defaultnights_toload)){
+					echo '*def03';
 					$defaultnights_toload[] = $defaultnights;
 				}